Bring PR #31776(Adjust storage for large payloads) from tcp branch to master (#33001)
* Adjust storage and processing in SystemPacketBuffer (#31776)

and associated code for large payloads.

Extend uint16_t type variables to size_t for the APIs
and all applicable places.

* Address review comments

+++ src/protocols/secure_channel/CASESession.cpp
@@ -1582,8 +1582,8 @@ CHIP_ERROR CASESession::HandleSigma3a(System::PacketBufferHandle && msg)
     TLV::TLVReader decryptedDataTlvReader;
     TLV::TLVType containerType = TLV::kTLVType_Structure;
 
-    const uint8_t * buf   = msg->Start();
-    const uint16_t bufLen = msg->DataLength();
+    const uint8_t * buf = msg->Start();
+    const size_t bufLen = msg->DataLength();
 
     constexpr size_t kCaseOverheadForFutureTbeData = 128;
